Move safe_rmtree().
